+++ crates/cheatnet/src/cheatcodes/mod.rs
@@ -17,6 +17,7 @@ pub mod mock_call;
 pub mod prank;
 pub mod precalculate_address;
 pub mod roll;
+pub mod elect;
 pub mod spoof;
 pub mod spy_events;
 pub mod warp;