+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Forge
 
+#### Added
+
+- `elect` cheatcode for mocking the sequencer address
+
 #### Removed
 
 - `available_gas` attribute, it didn't compute correctly gas usage. Contract functions execution cost would not be included.
@@ -18,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Cast
 
 #### Fixed
+
 - scarb metadata in declare subcommand now takes manifest path from cli if passed instead of looking for it
 
 ## [0.10.0] - 2023-11-08